GEODE-5212: Convert ExportLogsIntegrationTest to DistributedTest
  Due to differences on how defaultDirectory is initialized between
  windows and unix causes this test to fail if its integration test.
  So converted it to be a DistributedTest.
